% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referentiels.R
\name{capacites}
\alias{capacites}
\title{Capacités d'un programme}
\usage{
capacites(niveau_id = NULL, discipline_id = "MAT", version_id = NULL)
}
\arguments{
\item{niveau_id}{Niveau scolaire facultatif (`6E`, `5E`, `2GT`, ...).}

\item{discipline_id}{Discipline, `MAT` par défaut.}

\item{version_id}{Version scolaire facultative, par exemple `2026_2027`.}
}
\description{
Capacités d'un programme
}
