% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{carte_concept_math}
\alias{carte_concept_math}
\title{Carte pedagogique d'un concept mathematique}
\usage{
carte_concept_math(concept_id)
}
\arguments{
\item{concept_id}{Identifiant du concept.}
}
\description{
Carte pedagogique d'un concept mathematique
}
