% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charte.R
\name{charte_eduschool}
\alias{charte_eduschool}
\title{Charte graphique eduschool}
\usage{
charte_eduschool()
}
\value{
Un data.frame de couleurs nommees.
}
\description{
Retourne les couleurs d'identite utilisees par les fiches et graphiques.
}
