% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si.R
\name{colonnes_si}
\alias{colonnes_si}
\title{Colonnes du mini-SI eduschool}
\usage{
colonnes_si()
}
\value{
Un data.frame.
}
\description{
Retourne le contrat de colonnes : type de stockage, type semantique et role.
}
