% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{composition_exercice_math}
\alias{composition_exercice_math}
\title{Composition pedagogique d'un type d'exercice}
\usage{
composition_exercice_math(type_exercice_id)
}
\arguments{
\item{type_exercice_id}{Identifiant du type d'exercice.}
}
\description{
Composition pedagogique d'un type d'exercice
}
