% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{concepts_exercices_math}
\alias{concepts_exercices_math}
\title{Concepts mobilises par les types d'exercices mathematiques}
\usage{
concepts_exercices_math(type_exercice_id = NULL, concept_id = NULL)
}
\arguments{
\item{type_exercice_id}{Type d'exercice facultatif.}

\item{concept_id}{Concept facultatif.}
}
\description{
Concepts mobilises par les types d'exercices mathematiques
}
