% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{concepts_math}
\alias{concepts_math}
\title{Concepts mathematiques}
\usage{
concepts_math(domaine = NULL, statut = NULL, niveau_introduction = NULL)
}
\arguments{
\item{domaine}{Domaine facultatif.}

\item{statut}{Statut pedagogique facultatif.}

\item{niveau_introduction}{Niveau d'introduction facultatif.}
}
\description{
Concepts mathematiques
}
