% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charte.R
\name{couleur_cycle}
\alias{couleur_cycle}
\title{Couleur d'un cycle scolaire}
\usage{
couleur_cycle(cycle_id)
}
\arguments{
\item{cycle_id}{Identifiant de cycle (`"C3"`, `"C4"`,
`"CYCLE_TERMINAL"`) ou `"LYCEE"` pour la seconde.}
}
\value{
Une couleur au format hexadecimal.
}
\description{
Couleur d'un cycle scolaire
}
