% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{couverture_documentation}
\alias{couverture_documentation}
\title{Couverture de la documentation}
\usage{
couverture_documentation(niveau = NULL)
}
\arguments{
\item{niveau}{Niveau facultatif.}
}
\description{
Couverture de la documentation
}
