% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees_reelles.R
\name{decrire_source}
\alias{decrire_source}
\title{Decrire une source de donnees}
\usage{
decrire_source(source)
}
\arguments{
\item{source}{Objet produit par [source_insee_melodi()].}
}
\value{
Un data.frame a une ligne decrivant la source.
}
\description{
Decrire une source de donnees
}
