% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembles_nombres.R
\name{diagramme_ensembles_nombres}
\alias{diagramme_ensembles_nombres}
\title{Diagramme des ensembles de nombres}
\usage{
diagramme_ensembles_nombres()
}
\value{
Un objet ggplot.
}
\description{
Produit avec ggplot2 un schema de cadres emboites montrant les inclusions
N dans Z, Z dans D, D dans Q et Q dans R. Des exemples sont places dans
la zone propre a chaque ensemble pour montrer ce que chaque ensemble ajoute.
}
