% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrammes.R
\name{diagramme_package}
\alias{diagramme_package}
\title{Produire un diagramme technique du package}
\usage{
diagramme_package(
  type = c("architecture_si", "tests_package"),
  fichier = NULL,
  ouvrir = FALSE
)
}
\arguments{
\item{type}{Diagramme technique : `"architecture_si"` ou
`"tests_package"`.}

\item{fichier}{Chemin du fichier HTML à produire. Si `NULL`, un fichier
temporaire est cree.}

\item{ouvrir}{Ouvrir le fichier avec l'application associée après sa création.}
}
\value{
Invisiblement, le chemin absolu du fichier HTML produit.
}
\description{
Produire un diagramme technique du package
}
