% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrammes.R
\name{diagramme_parcours_scolaire}
\alias{diagramme_parcours_scolaire}
\title{Produire le diagramme des parcours scolaires}
\usage{
diagramme_parcours_scolaire(fichier = NULL, ouvrir = FALSE)
}
\arguments{
\item{fichier}{Chemin du fichier HTML à produire. Si `NULL`, un fichier
temporaire est cree.}

\item{ouvrir}{Ouvrir le fichier avec l'application associée après sa création.}
}
\value{
Invisiblement, le chemin absolu du fichier HTML produit.
}
\description{
Raccourci vers `produire_diagramme_html("parcours_scolaire", ...)`.
}
