% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrammes.R
\name{diagrammes_disponibles}
\alias{diagrammes_disponibles}
\title{Diagrammes disponibles}
\usage{
diagrammes_disponibles()
}
\value{
Un data.frame décrivant les diagrammes disponibles.
}
\description{
Retourne le catalogue des diagrammes fournis par eduschool. Les mêmes
modèles alimentent les sorties HTML et SVG.
}
