% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eduschool.R
\name{eduschool}
\alias{eduschool}
\title{Decouvrir eduschool}
\usage{
eduschool()
}
\value{
Invisiblement, une liste des fonctions par usage.
}
\description{
Presente les principales portes d'entree du package.
Les ressources scolaires historiques ne garantissent pas l'actualite
des programmes officiels.
}
