% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{eduschool_init}
\alias{eduschool_init}
\title{Initialiser un contexte eduschool}
\usage{
eduschool_init(fichier_base = ":memory:", importer = TRUE)
}
\arguments{
\item{fichier_base}{Base DuckDB, en mémoire par défaut.}

\item{importer}{Importer les CSV dans DuckDB.}
}
\description{
Initialiser un contexte eduschool
}
