% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{enseignements_optionnels_lycee}
\alias{enseignements_optionnels_lycee}
\title{Enseignements optionnels au lycee general}
\usage{
enseignements_optionnels_lycee(niveau_id = "1G", version_id = "2026_2027")
}
\arguments{
\item{niveau_id}{`"1G"` ou `"TG"`.}

\item{version_id}{Version scolaire. Par defaut `"2026_2027"`.}
}
\value{
Un data.frame avec libelle et volume horaire.
}
\description{
Les enseignements optionnels completent le parcours mais ne doivent pas etre
confondus avec les enseignements de specialite.
}
