% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembles_nombres.R
\name{ensembles_nombres}
\alias{ensembles_nombres}
\title{Ensembles de nombres usuels}
\usage{
ensembles_nombres()
}
\value{
Un data.frame ordonne de N a R.
}
\description{
Retourne la petite table pedagogique utilisee par les fiches, schemas et
exercices consacres aux ensembles de nombres en seconde.
}
