% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{erreurs_math}
\alias{erreurs_math}
\title{Erreurs frequentes en mathematiques}
\usage{
erreurs_math(concept_id = NULL, niveau_id = NULL)
}
\arguments{
\item{concept_id}{Concept facultatif.}

\item{niveau_id}{Niveau facultatif.}
}
\description{
Erreurs frequentes en mathematiques
}
