% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examens.R
\name{examen}
\alias{examen}
\title{Decrire un examen}
\usage{
examen(code, session)
}
\arguments{
\item{code}{Code de l'examen.}

\item{session}{Session de l'examen.}
}
\value{
Une ligne du referentiel des examens.
}
\description{
Decrire un examen
}
