% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examens.R
\name{examens}
\alias{examens}
\title{Examens modelises}
\usage{
examens(code = NULL, session = NULL)
}
\arguments{
\item{code}{Code d'examen facultatif, par exemple `"DNB"`.}

\item{session}{Session facultative.}
}
\value{
Un data.frame.
}
\description{
Examens modelises
}
