% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{familles_revision}
\alias{familles_revision}
\title{Familles de fiches de revision}
\usage{
familles_revision(niveau_id = NULL)
}
\arguments{
\item{niveau_id}{Niveau scolaire facultatif. Lorsque renseigne, seules les
familles possedant une fiche pour ce niveau sont retournees.}
}
\value{
Un data.frame.
}
\description{
Familles de fiches de revision
}
