% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contenus.R
\name{fiches}
\alias{fiches}
\title{Lister les fiches de revision disponibles}
\usage{
fiches(niveau = NULL)
}
\arguments{
\item{niveau}{Niveau facultatif, par exemple "6E" ou "2GT".}
}
\value{
Un data.frame avec niveau, fiche et fichier.
}
\description{
Lister les fiches de revision disponibles
}
