% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{filieres_postbac}
\alias{filieres_postbac}
\title{Grandes filieres post-bac}
\usage{
filieres_postbac(categorie = NULL)
}
\arguments{
\item{categorie}{Categorie facultative : `UNIVERSITE`, `LYCEE` ou `ECOLE`.}
}
\value{
Un data.frame.
}
\description{
Grandes filieres post-bac
}
