% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formes_questions.R
\name{formes_questions}
\alias{formes_questions}
\title{Formes de questions}
\usage{
formes_questions()
}
\value{
Un data.frame avec les colonnes `notion`, `forme` et `explication`.
}
\description{
Lit le catalogue pedagogique des formes de questions recensees dans eduschool.
}
