% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{formules_math}
\alias{formules_math}
\title{Formules mathematiques essentielles}
\usage{
formules_math(concept_id = NULL, niveau_id = NULL)
}
\arguments{
\item{concept_id}{Concept facultatif.}

\item{niveau_id}{Niveau facultatif.}
}
\description{
Formules mathematiques essentielles
}
