% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabarits_examens.R
\name{gabarit_examen}
\alias{gabarit_examen}
\title{Decrire un gabarit d'exercice}
\usage{
gabarit_examen(gabarit_id)
}
\arguments{
\item{gabarit_id}{Identifiant du gabarit.}
}
\value{
Une liste contenant le gabarit, ses concepts et ses parametres.
}
\description{
Decrire un gabarit d'exercice
}
