% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercices_composes.R
\name{gabarit_exercice_compose}
\alias{gabarit_exercice_compose}
\title{Decrire un gabarit d'exercice compose}
\usage{
gabarit_exercice_compose(gabarit_compose_id)
}
\arguments{
\item{gabarit_compose_id}{Identifiant du gabarit compose.}
}
\value{
Une liste avec gabarit, questions et ressources.
}
\description{
Decrire un gabarit d'exercice compose
}
