% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercices_composes.R
\name{gabarits_exercices_composes}
\alias{gabarits_exercices_composes}
\title{Gabarits d'exercices composes}
\usage{
gabarits_exercices_composes(
  examen_code = NULL,
  partie_type = NULL,
  statut = "ACTIF"
)
}
\arguments{
\item{examen_code}{Code d'examen facultatif.}

\item{partie_type}{Type de partie facultatif.}

\item{statut}{Statut des gabarits. Par defaut, seuls les actifs sont retournes.}
}
\value{
Un data.frame.
}
\description{
Gabarits d'exercices composes
}
