% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{generer_essentiel}
\alias{generer_essentiel}
\title{Generer la fiche essentielle d'un niveau}
\usage{
generer_essentiel(niveau_id)
}
\arguments{
\item{niveau_id}{Identifiant du niveau, par exemple `"2GT"`.}
}
\value{
Un objet `eduschool_revision` tres synthetique.
}
\description{
Generer la fiche essentielle d'un niveau
}
