% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabarits_examens.R
\name{generer_gabarit_examen}
\alias{generer_gabarit_examen}
\title{Instancier un gabarit d'examen}
\usage{
generer_gabarit_examen(gabarit_id, seed = NULL)
}
\arguments{
\item{gabarit_id}{Identifiant du gabarit.}

\item{seed}{Graine aleatoire facultative.}
}
\value{
Une liste avec metadonnees, enonce, reponse et parametres tires.
}
\description{
Genere un enonce et sa reponse a partir d'un gabarit parametrique. Dans un
meme contexte de generation, la graine permet de retrouver les memes tirages
pseudo-aleatoires.
}
