% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{graphique}
\alias{graphique}
\title{Dessiner un graphique pedagogique reutilisable}
\usage{
graphique(id, ...)
}
\arguments{
\item{id}{Identifiant unique du graphique.}

\item{...}{Parametres de la fonction graphique.}
}
\description{
Dessiner un graphique pedagogique reutilisable
}
