% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheses.R
\name{horaires_niveau}
\alias{horaires_niveau}
\title{Horaires d'un niveau scolaire}
\usage{
horaires_niveau(niveau_id, version_id = "2026_2027")
}
\arguments{
\item{niveau_id}{Identifiant du niveau, par exemple `6E`.}

\item{version_id}{Version scolaire, par exemple `2026_2027`.}
}
\description{
Horaires d'un niveau scolaire
}
