% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{illustrer_question}
\alias{illustrer_question}
\title{Dessiner une illustration de question}
\usage{
illustrer_question(q)
}
\arguments{
\item{q}{Question instanciee.}
}
\description{
Dessiner une illustration de question
}
