% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapports.R
\name{libelle_capacite}
\alias{libelle_capacite}
\title{Libelle d'une capacite de programme}
\usage{
libelle_capacite(capacite_id)
}
\arguments{
\item{capacite_id}{Identifiant de la capacite ou de l'item de programme.}
}
\value{
Une chaine de caracteres contenant le libelle de la capacite, ou
  `NULL` si `capacite_id` est absent.
}
\description{
Retourne le libelle associe a un identifiant de capacite ou d'item de
programme. Si le referentiel n'est pas disponible ou si l'identifiant
n'est pas trouve, l'identifiant fourni est retourne tel quel.
}
