% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapports.R
\name{libelle_niveau}
\alias{libelle_niveau}
\title{Libelle d'un niveau scolaire}
\usage{
libelle_niveau(niveau_id)
}
\arguments{
\item{niveau_id}{Identifiant du niveau scolaire.}
}
\value{
Une chaine de caracteres contenant le libelle du niveau.
}
\description{
Retourne le libelle associe a un identifiant de niveau. Si le referentiel
n'est pas disponible ou si l'identifiant n'est pas trouve, l'identifiant
fourni est retourne tel quel.
}
