% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{methodes_exercices_math}
\alias{methodes_exercices_math}
\title{Methodes mobilisees par les types d'exercices mathematiques}
\usage{
methodes_exercices_math(type_exercice_id = NULL, methode_id = NULL)
}
\arguments{
\item{type_exercice_id}{Type d'exercice facultatif.}

\item{methode_id}{Methode facultative.}
}
\description{
Methodes mobilisees par les types d'exercices mathematiques
}
