% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{methodes_math}
\alias{methodes_math}
\title{Methodes mathematiques}
\usage{
methodes_math(concept_id = NULL, niveau_id = NULL, domaine = NULL)
}
\arguments{
\item{concept_id}{Concept facultatif.}

\item{niveau_id}{Niveau facultatif.}

\item{domaine}{Domaine facultatif.}
}
\description{
Methodes mathematiques
}
