% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{notions_capacite}
\alias{notions_capacite}
\title{Notions associées à une capacité}
\usage{
notions_capacite(capacite_id)
}
\arguments{
\item{capacite_id}{Identifiant(s) de capacité.}
}
\description{
Notions associées à une capacité
}
