% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{obtenir_rappel}
\alias{obtenir_rappel}
\title{Lire le rappel pédagogique d'une notion}
\usage{
obtenir_rappel(notion_id, collapse = "\\n")
}
\arguments{
\item{notion_id}{Identifiant de notion.}

\item{collapse}{Séparateur des lignes.}
}
\description{
Lire le rappel pédagogique d'une notion
}
