% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decouvrir.R
\name{parcours}
\alias{parcours}
\title{Parcours scolaire}
\usage{
parcours(niveau = NULL)
}
\arguments{
\item{niveau}{Niveau scolaire facultatif, par exemple "3E" ou "2GT".}
}
\value{
Une liste structuree contenant les niveaux, les noeuds et les liens utiles.
}
\description{
Construit les informations de parcours scolaire representees dans eduschool.
}
