% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{parcoursup_calendrier}
\alias{parcoursup_calendrier}
\title{Jalons dates d'une campagne Parcoursup}
\usage{
parcoursup_calendrier(campagne_id = "PS2026")
}
\arguments{
\item{campagne_id}{Identifiant de campagne, par defaut `PS2026`.}
}
\value{
Un data.frame ordonne des jalons dates de la campagne.
}
\description{
Le calendrier est modelise sous forme evenementielle : ajouter un jalon a une
nouvelle campagne ne necessite donc aucune modification du schema.
}
