% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{parcoursup_etapes}
\alias{parcoursup_etapes}
\title{Etapes de Parcoursup}
\usage{
parcoursup_etapes()
}
\value{
Un data.frame ordonne des grandes etapes structurelles de la procedure.
}
\description{
Etapes de Parcoursup
}
