% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{parcoursup_nouveautes}
\alias{parcoursup_nouveautes}
\title{Nouveautes d'une campagne Parcoursup}
\usage{
parcoursup_nouveautes(campagne_id = "PS2026")
}
\arguments{
\item{campagne_id}{Identifiant de campagne, par defaut `PS2026`.}
}
\value{
Un data.frame ordonne.
}
\description{
Nouveautes d'une campagne Parcoursup
}
