% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_questions.R
\name{parser_questions}
\alias{parser_questions}
\title{Lire une banque de questions}
\usage{
parser_questions(fichier)
}
\arguments{
\item{fichier}{Fichier Markdown contenant les questions.}
}
\value{
Une liste de questions structurees.
}
\description{
Lire une banque de questions
}
