% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{plateformes_admission}
\alias{plateformes_admission}
\title{Plateformes nationales d'admission post-bac}
\usage{
plateformes_admission()
}
\value{
Une liste avec `plateformes` et `caracteristiques`.
}
\description{
Retourne APB et Parcoursup ainsi que leurs caracteristiques documentees.
}
