% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{prerequis_capacite}
\alias{prerequis_capacite}
\title{Prérequis des notions associées à une capacité}
\usage{
prerequis_capacite(capacite_id, recursif = FALSE)
}
\arguments{
\item{capacite_id}{Identifiant(s) de capacité.}

\item{recursif}{Inclure tous les prérequis transitifs.}
}
\description{
Prérequis des notions associées à une capacité
}
