% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{prerequis_notion}
\alias{prerequis_notion}
\title{Prérequis d'une notion}
\usage{
prerequis_notion(notion_id, recursif = FALSE)
}
\arguments{
\item{notion_id}{Identifiant(s) de notion.}

\item{recursif}{Inclure tous les prérequis transitifs.}
}
\description{
Prérequis d'une notion
}
