% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.eduschool_context}
\alias{print.eduschool_context}
\title{Afficher un contexte eduschool}
\usage{
\method{print}{eduschool_context}(x, ...)
}
\arguments{
\item{x}{Contexte eduschool.}

\item{...}{Arguments supplémentaires.}
}
\description{
Afficher un contexte eduschool
}
