% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_math.R
\name{produire_carte_math}
\alias{produire_carte_math}
\title{Produire la carte des mathematiques}
\usage{
produire_carte_math(
  carte = carte_math(),
  fichier = NULL,
  largeur = 10,
  hauteur = 7,
  ouvrir = TRUE
)
}
\arguments{
\item{carte}{Objet produit par [carte_math()].}

\item{fichier}{Fichier PNG a produire. Si `NULL`, utilise
`eduschool-carte-math.png` dans le repertoire courant.}

\item{largeur}{Largeur du rendu en pouces.}

\item{hauteur}{Hauteur du rendu en pouces.}

\item{ouvrir}{Ouvrir le rendu apres sa creation.}
}
\value{
Invisiblement, le chemin absolu du fichier produit.
}
\description{
Produire la carte des mathematiques
}
