% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{produire_schema_parcoursup_svg}
\alias{produire_schema_parcoursup_svg}
\title{Produire le schema des etapes Parcoursup}
\usage{
produire_schema_parcoursup_svg(fichier = NULL)
}
\arguments{
\item{fichier}{Chemin du SVG. Si `NULL`, utilise un fichier temporaire.}
}
\value{
Invisiblement, le chemin absolu du SVG produit.
}
\description{
Produire le schema des etapes Parcoursup
}
