% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referentiels.R
\name{programmes}
\alias{programmes}
\title{Programmes}
\usage{
programmes(discipline_id = NULL, niveau_id = NULL)
}
\arguments{
\item{discipline_id}{Identifiant de discipline, par exemple `MAT`.}

\item{niveau_id}{Identifiant de niveau facultatif.}
}
\description{
Programmes
}
