% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{questions}
\alias{questions}
\title{Lire les questions d'une notion pilote}
\usage{
questions(notion, niveau = "")
}
\arguments{
\item{notion}{Identifiant d'une notion ou d'une famille.}

\item{niveau}{Niveau scolaire facultatif pour filtrer les notions d'une famille.}
}
\description{
Lire les questions d'une notion pilote
}
