% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contenus.R
\name{quiz}
\alias{quiz}
\title{Tirer un quiz dans la banque de questions}
\usage{
quiz(
  n = 5L,
  niveau = NULL,
  theme = NULL,
  notion = NULL,
  type = NULL,
  banque = NULL
)
}
\arguments{
\item{n}{Nombre de questions.}

\item{niveau}{Filtre facultatif sur le niveau scolaire.}

\item{theme}{Filtre facultatif sur le theme mathematique.}

\item{notion}{Filtre facultatif.}

\item{type}{Filtre facultatif.}

\item{banque}{Filtre facultatif.}
}
\value{
Une liste de questions structurees.
}
\description{
Tirer un quiz dans la banque de questions
}
