% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{rappels_capacite}
\alias{rappels_capacite}
\title{Rappels associés à une capacité}
\usage{
rappels_capacite(capacite_id)
}
\arguments{
\item{capacite_id}{Identifiant de capacité.}
}
\description{
Rappels associés à une capacité
}
