\name{rapports-exercices}
\alias{produire_fiche_exercices}
\alias{produire_corrige_exercices}
\alias{produire_rapport_exercices}
\alias{libelle_niveau}
\alias{libelle_capacite}
\title{Produire des fiches et corrigés}
\description{Fonctions de sortie LaTeX/PDF pour les lots d'exercices et fonctions utilitaires de libellé.}
\details{La compilation PDF nécessite un exécutable \code{pdflatex}. Sans celui-ci, le fichier TeX peut néanmoins être produit.}
