\name{referentiels}
\alias{niveaux}
\alias{voies}
\alias{series}
\alias{disciplines}
\alias{enseignements}
\title{Consulter les référentiels}
\usage{
niveaux()
voies()
series()
disciplines()
enseignements()
}
\description{Retourne les principaux référentiels distribués avec le package sous forme de data.frame.}
\value{Un \code{data.frame}.}
