% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{relations_concepts_math}
\alias{relations_concepts_math}
\title{Relations entre concepts mathematiques}
\usage{
relations_concepts_math(concept_id = NULL, type_relation = NULL)
}
\arguments{
\item{concept_id}{Concept facultatif.}

\item{type_relation}{Type de relation facultatif.}
}
\description{
Relations entre concepts mathematiques
}
