% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si.R
\name{relations_si}
\alias{relations_si}
\title{Relations du mini-SI eduschool}
\usage{
relations_si()
}
\value{
Un data.frame.
}
\description{
Retourne les relations declarees entre cles etrangeres et tables de reference.
}
