% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si.R
\name{resume_controles_si}
\alias{resume_controles_si}
\title{Resume des controles d'integrite}
\usage{
resume_controles_si()
}
\value{
Un data.frame agrege par type de controle.
}
\description{
Resume des controles d'integrite
}
