% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths-ressources.R
\name{series}
\alias{series}
\title{Series du lycee (referentiel historique)}
\usage{
series()
}
\value{
Un data.frame. Donnees reprises de la sauvegarde v1.
}
\description{
Series du lycee (referentiel historique)
}
