% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{series_technologiques}
\alias{series_technologiques}
\title{Series technologiques disponibles}
\usage{
series_technologiques()
}
\value{
Un data.frame des series de la voie technologique.
}
\description{
Series technologiques disponibles
}
