% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{specialites_generales}
\alias{specialites_generales}
\title{Enseignements de specialite de la voie generale}
\usage{
specialites_generales()
}
\value{
Un data.frame des enseignements de type `SPECIALITE`.
}
\description{
Enseignements de specialite de la voie generale
}
