% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examens.R
\name{structure_examen}
\alias{structure_examen}
\title{Structure d'un examen}
\usage{
structure_examen(code, session)
}
\arguments{
\item{code}{Code de l'examen.}

\item{session}{Session de l'examen.}
}
\value{
Une liste contenant `examen` et `parties`.
}
\description{
Retourne l'examen et ses parties officielles modelisees dans eduschool.
}
