% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{types_exercices_math}
\alias{types_exercices_math}
\title{Types d'exercices mathematiques}
\usage{
types_exercices_math(concept_id = NULL, niveau_id = NULL, domaine = NULL)
}
\arguments{
\item{concept_id}{Concept facultatif.}

\item{niveau_id}{Niveau facultatif.}

\item{domaine}{Domaine facultatif.}
}
\description{
Types d'exercices mathematiques
}
