% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economie.R
\name{variation_pourcentage}
\alias{variation_pourcentage}
\title{Calculer une variation en pourcentage}
\usage{
variation_pourcentage(valeur_initiale, valeur_finale)
}
\arguments{
\item{valeur_initiale}{Valeur de depart.}

\item{valeur_finale}{Valeur d'arrivee.}
}
\value{
La variation en pourcentage.
}
\description{
Calculer une variation en pourcentage
}
