% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths-ressources.R
\name{voies}
\alias{voies}
\title{Voies scolaires (referentiel historique)}
\usage{
voies()
}
\value{
Un data.frame. Donnees reprises de la sauvegarde v1.
}
\description{
Voies scolaires (referentiel historique)
}
