% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{chercher_notions}
\alias{chercher_notions}
\title{Rechercher des notions}
\usage{
chercher_notions(texte, discipline_id = "MAT")
}
\arguments{
\item{texte}{Texte ou fragments à rechercher.}

\item{discipline_id}{Discipline, `MAT` par défaut.}
}
\description{
Rechercher des notions
}
