% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eduschool-package.R
\docType{package}
\name{eduschool-package}
\alias{eduschool}
\alias{eduschool-package}
\title{eduschool : parcours scolaire et mathematiques}
\description{
`eduschool` cartographie la scolarite des collegiens et lyceens en France et
propose des outils pour reviser et s'entrainer en mathematiques. Son
architecture de donnees reste generique et peut accueillir d'autres
disciplines, sans que cela constitue l'objectif principal du projet.
}
\details{
L'API principale s'organise autour de quelques verbes : [parcours()],
[orientation()], [programme()], [revision()] et [exercices()]. Les fonctions
historiques et les tables du mini-SI restent disponibles pour les usages
avances, la maintenance et la contribution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gilles13.github.io/eduschool/}
}

}
\author{
\strong{Maintainer}: Gilles FIDANI \email{gillesfidani@gmail.com}

Authors:
\itemize{
  \item Gilles FIDANI \email{gillesfidani@gmail.com}
}

}
\keyword{internal}
