\name{eduschool-package}
\alias{eduschool-package}
\alias{eduschool}
\title{eduschool : référentiels scolaires et outils de révision}
\description{Package R pour consulter des référentiels scolaires français, relier programmes, capacités et notions pédagogiques, et générer des exercices reproductibles.}
\details{Les ressources persistantes sont distribuées sous forme de CSV et Markdown. Les consultations simples n'exigent pas de connexion DuckDB.}
\docType{package}
