% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eduschool-package.R
\docType{package}
\name{eduschool-package}
\alias{eduschool}
\alias{eduschool-package}
\title{eduschool : référentiels scolaires et révisions}
\description{
`eduschool` organise des référentiels scolaires français, une base
documentaire pédagogique et des générateurs d'exercices reproductibles.
Les consultations simples utilisent les CSV installés avec le package ;
DuckDB est disponible pour les usages relationnels plus avancés.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gilles13.github.io/eduschool/}
}

}
\author{
\strong{Maintainer}: Gilles FIDANI \email{gillesfidani@gmail.com}

Authors:
\itemize{
  \item Gilles FIDANI \email{gillesfidani@gmail.com}
}

}
\keyword{internal}
