% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decouvrir.R
\name{eduschool}
\alias{eduschool}
\title{Decouvrir eduschool}
\usage{
eduschool()
}
\value{
Invisiblement, le nom des fonctions presentees.
}
\description{
Affiche les principales portes d'entree du package.
}
\examples{
eduschool()
}
