% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{eduschool_path}
\alias{eduschool_path}
\title{Chemin vers une ressource installée}
\usage{
eduschool_path(..., must_work = TRUE)
}
\arguments{
\item{...}{Composantes du chemin à l'intérieur de `inst/`.}

\item{must_work}{Erreur si la ressource n'existe pas.}
}
\description{
Retourne le chemin d'une ressource distribuée avec eduschool. Cette
fonction fonctionne aussi depuis un arbre source non encore installé.
}
