\name{gabarit_examen}
\alias{gabarit_examen}
\title{Decrire un gabarit d'exercice}
\usage{gabarit_examen(gabarit_id)}
\arguments{\item{gabarit_id}{Identifiant du gabarit.}}
\value{Une liste contenant le gabarit, ses concepts et ses parametres.}
\description{Retourne la definition complete d'un gabarit d'examen.}
