% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheses.R
\name{horaires_niveau}
\alias{horaires_niveau}
\title{Horaires d'un niveau scolaire}
\usage{
horaires_niveau(niveau_id, version_id = "2026_2027", serie_id = NULL)
}
\arguments{
\item{niveau_id}{Identifiant du niveau, par exemple `6E`.}

\item{version_id}{Version scolaire, par exemple `2026_2027`.}

\item{serie_id}{Serie technologique ou generale facultative. Lorsqu'elle est
absente, seuls les horaires communs au niveau sont retournes. Pour un niveau
dont la grille est entierement definie par serie (par exemple `1T` ou `TT`),
`serie_id` est obligatoire.}
}
\description{
Horaires d'un niveau scolaire
}
