% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{illustrer_question}
\alias{illustrer_question}
\title{Dessiner le triangle des questions de Pythagore}
\usage{
illustrer_question(q)
}
\arguments{
\item{q}{Question instanciée.}
}
\description{
Dessiner le triangle des questions de Pythagore
}
