% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decouvrir.R
\name{notions}
\alias{notions}
\title{Notions mathematiques}
\usage{
notions(niveau = NULL, theme = NULL)
}
\arguments{
\item{niveau}{Niveau facultatif.}

\item{theme}{Theme facultatif.}
}
\value{
Une liste structuree contenant les notions et leur organisation par theme.
}
\description{
Construit les notions connues d'eduschool, organisees par niveau et par theme.
}
