% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheses.R
\name{notions_niveau}
\alias{notions_niveau}
\title{Notions documentées pour un niveau}
\usage{
notions_niveau(niveau_id, discipline_id = NULL, version_id = "2026_2027")
}
\arguments{
\item{niveau_id}{Identifiant du niveau, par exemple `6E`.}

\item{discipline_id}{Discipline facultative.}

\item{version_id}{Version scolaire, par exemple `2026_2027`.}
}
\description{
Notions documentées pour un niveau
}
