% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{question}
\alias{question}
\title{Instancier une question finie}
\usage{
question(definition)
}
\arguments{
\item{definition}{Element de questions(notion)$questions.}
}
\description{
Instancier une question finie
}
