% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{question}
\alias{question}
\title{Instancier une question (JSON local de confiance uniquement)}
\usage{
question(definition)
}
\arguments{
\item{definition}{Élément de questions(notion)$questions.}
}
\description{
Instancier une question (JSON local de confiance uniquement)
}
