% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contenus.R
\name{question}
\alias{question}
\title{Tirer une formulation de question}
\usage{
question(
  niveau = NULL,
  theme = NULL,
  notion = NULL,
  type = NULL,
  banque = NULL
)
}
\arguments{
\item{niveau}{Filtre facultatif sur le niveau scolaire.}

\item{theme}{Filtre facultatif sur le theme mathematique.}

\item{notion}{Filtre facultatif.}

\item{type}{Filtre facultatif.}

\item{banque}{Filtre facultatif.}
}
\value{
Une question structuree.
}
\description{
Tirer une formulation de question
}
