% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematiques.R
\name{questions}
\alias{questions}
\title{Lire les questions d'une notion pilote}
\usage{
questions(notion, niveau = "")
}
\arguments{
\item{notion}{"addition_fractions" ou "pythagore".}
}
\description{
Lire les questions d'une notion pilote
}
