% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redaction_examens.R
\name{rediger_examen}
\alias{rediger_examen}
\title{Rediger une partie d'un examen compose}
\usage{
rediger_examen(sujet, partie = 1)
}
\arguments{
\item{sujet}{Objet produit par `composer_examen()`.}

\item{partie}{Numero d'ordre ou identifiant de la partie a rediger.}
}
\value{
Un objet `eduschool_examen_redige`.
}
\description{
Transforme un squelette produit par `composer_examen()` en objet intermediaire
contenant les enonces, reponses, corrections et specifications de ressources
graphiques. Cette fonction ne produit pas encore de document PDF.
}
