% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_multiplication.R
\name{table_multiplication}
\alias{table_multiplication}
\title{Tables de multiplication}
\usage{
table_multiplication(n = 1:9)
}
\arguments{
\item{n}{Tables a produire. Par defaut, de 1 a 9.}
}
\description{
Tables de multiplication
}
